MD-548: Remove reference to mapr-release pom.
* Bumped the Drill+plugin version to 1.6.0-SNAPSHOT.
* Updated MapR client artifacts to 5.1.0.
* Added missing Apache license headers to few files.
* Added a base test class BaseJsonTest for all OJAI Unit tests.
* Invoke GuavaPatcher from BaseJsonTest.
